Changed vissprite->mo->lightlevel to (byte)

+++ doomsday/Include/r_things.h
@@ -67,7 +67,7 @@ typedef struct vissprite_s {
             float           floorglowamount;
             float           ceilglowamount;
             byte            rgb[3]; // Sector light color.
-            int             lightlevel;
+            byte            lightlevel;
             float           alpha;
             float           visoff[3];  // Last-minute offset to coords.
             struct modeldef_s *mf, *nextmf;
